Fix repair projection ownership
Problem: asynchronous count repair could rewrite recovery projection fields and publish a partial topic snapshot before directory reconciliation.

Root cause: repair updated recovery_copy and topic aggregates outside the directory lock, so it could interleave with a parsed reconcile snapshot.

Fix: serialize repair with the directory lock, update only source-owned fields, preserve repaired source state across legacy sidecars, and leave projection publication to one directory transaction. Extract upsert and repair ownership into focused files to satisfy repository size gates.

Verification: go test ./...; go test -race ./internal/sessioncatalog; go vet ./internal/sessioncatalog; go run ./tools/repolint; BenchmarkReconcileDirectory10K (3x, 2.74s/op).
